Adding isStalemate().

+++ src/main.cc
@@ -20,7 +20,7 @@ int main() {
   
   string command, arg1, arg2;
   char piece;
-  int blackPoints, whitePoints;
+  float blackPoints, whitePoints;
   Piece* pieces = new Blank;
   TextDisplay* textDisplay;
